refactor: move useDetectNativeToken to modules/trade
