+++ src/modules/swap/hooks/useIsSwapEth.ts
@@ -1,5 +1,4 @@
-import { useDetectNativeToken } from './useDetectNativeToken'
-
+import { useDetectNativeToken } from '../../trade/hooks/useDetectNativeToken'
 import { useIsWrapOrUnwrap } from '../../trade/hooks/useIsWrapOrUnwrap'
 
 export function useIsSwapEth(): boolean {
