+++ src/modules/trade/hooks/useDetectNativeToken.ts
@@ -6,7 +6,7 @@ import { Token } from '@uniswap/sdk-core'
 import WXDAI_LOGO_URI from 'legacy/assets/cow-swap/wxdai.png'
 import { WETH_LOGO_URI } from 'legacy/constants'
 import { DEFAULT_NETWORK_FOR_LISTS } from 'legacy/constants/lists'
-import { WRAPPED_NATIVE_CURRENCY as WETH, GpEther as ETHER } from 'legacy/constants/tokens'
+import { GpEther as ETHER, WRAPPED_NATIVE_CURRENCY as WETH } from 'legacy/constants/tokens'
 import { supportedChainId } from 'legacy/utils/supportedChainId'
 
 import { useTradeState } from 'modules/trade/hooks/useTradeState'
@@ -14,7 +14,6 @@ import { useWalletInfo } from 'modules/wallet'
 
 import { useTokenBySymbolOrAddress } from 'common/hooks/useTokenBySymbolOrAddress'
 
-// TODO: move it to `modules/trade`
 export function useDetectNativeToken() {
   const { chainId } = useWalletInfo()
   const { state } = useTradeState()
